Resolves: tdf#112258 correctly pick items from AutoFilter selection
Combining only one child with one parent to lookup is not sufficient if we have
year,month,day chains. The entire chain is needed.

Regression from

    commit 9b70798b794ed7d7052d2cf497098a30fb9dcb75
    Date:   Thu Jun 1 18:27:13 2017 +0200

        tdf#108259 Fix nested checkbox handling in autofilter popup

and

    commit dc3cea57ee53fddc2a1f70eb02e14d4652973f43
    Date:   Wed May 31 10:59:42 2017 +0200

        tdf#108259 Enable autofilter with many different values

which combined landed in 5.4 as

    commit be55eeb81492bfaa3300a0d6815e0ff28563562b
    Date:   Wed May 31 10:59:42 2017 +0200

        tdf#108259 Enable autofilter with many different values

This on the other hand makes the change of these doubtful, as for each leaf
item the concatenation and the lookup has to be done. Something to be
investigated.

Change-Id: I07fd6367bb8da2adab94a43c45fe88391179f496