tdf#112289 OSX fpicker: don't default to true
When LO requests a value for the currently unimplemented
GPG encryption checkbox, the Aqua file picker defaults to
a "uno::Any <<= true", which triggers the GPG key dialog.

This simply returns an empty uno::Any for values of unknown
controls, which is handled fine.

If something else relies on the current default value, it
must to be properly implemented.

Change-Id: I3637ca833d59e19416e25f3096bc90a756aa8840