Clean up SvxSwFramePosString::GetString
...after fede58e9a69f41d14e051b217ccc8485696aa1dc
"-Werror,-Wtautological-compare with latest clang".  Even with a compiler that
chose int as the underlying type of StringId, eId could only be negative after
UB has happened elsewhere, but nevertheless keep checking for it in the assert
(which is a macro in which -Wtautological-compare is apparently disabled with
recent Clang trunk at least).

Change-Id: I448c6d48e9f84ef84089f4735776b6a66599dc67