tdf#112446 ooxmlimport: Orient=NONE when distance is given
Prior to commit 2b2134408c31b370068bb461b1ca189843117cb3
the style could only contain the default of NONE. So when
a position was specified, it was always paired with
HoriOrient == NONE. So it never caused problems until
that commit when the Frame's style orientation started
overriding the unset paragraph default.

When a position is specified, that needs to be paired with an
orientation of NONE in order to take effect.

Change-Id: Iab0057810270ba708a8855c2ec6db291cef17cfb
Reviewed-on: https://gerrit.libreoffice.org/42499
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>