Revert "tdf#108932 - prevent textbox shrink from actually growing"
This reverts commit 56c47c85cf1b903e98dc280e19b602ca3947ee13 which
was a trial attemp to fix a regression from tdf#91260.

Not reverting my jenkins comment that "Actually, I think that
the regression commit ought to be fully reverted. It "fixes" such
a strange situation and seems to have a lot of unintended
repercussions. Just look at tdf#91260 and see all of the
additional patches."

Change-Id: I6852f6d494f48f68eefc39019c44f2cbd96ec5a5
Reviewed-on: https://gerrit.libreoffice.org/42827
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>