tdf#112012 EMF+ dashed lines: correct dash length
Since the first guess of the right transformation seemed to be wrong,
here is the next try. The pure float value is multiplied with the pen width.
This is way more intuitive for me and renders the EMF files from the
bugtracker exactly like in MSO.

See 0babbdf9c640ad93d38edaa66d410ccf7c76d877 for the previous patch.

Change-Id: I74e82a0f9800d950e0d989184765a53e520b48e6
Reviewed-on: https://gerrit.libreoffice.org/42810
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>