tdf#112679 sw: fix copying of fieldmarks
Aha, now we know that the reason for the defensive programming
in lcl_AssureFieldMarksSet() was that there are actually 2
different use-cases for it: usually a new mark is inserted,
so there are no dummy characters and they must be inserted.

However when copying text, the dummy characters are copied too,
so they must not be inserted, or we get duplicate fieldmarks.

This also reverts commit 3b26dd0ed6f9a3fed25ee740e929c44ed63ed39d
which fixed the problem only for CHECKBOX_FIELDMARK in a
different way.

(regression from 5d440aeb070697bc4cf8b5b47c91aab373213c79)

Change-Id: I3c99b8c6d720951655198e682018794337859373