Number scanner: accept fractional separators in an ISO 8601 date+time string
In a strict ISO 8601 date+time string with 'T' separator the
Time100SecSep separating seconds and fractional seconds can be
either '.' period or ',' comma, so accept those in all locales.

Not accepting '.' in all locales was the cause of tdf#100822
before code was changed to use sax::Converter::parseDateTime()
instead.

Change-Id: Ica676050b52b11da64afbac6feabb43d9e985bc4