tdf#31814 EMF/EMF+ implement dual mode
There can be a dual mode flag set in the EMF+ data. If set,
we should process either EMF or EMF+ data, because the contained
information should be equal (says the documentation).
The more advanced EMF+ renderer is chosen here.

The flag is read in the EMFPlusReader and blocks the reading of other
EMF records than HEADER, EOF and COMMENT.

Thanks to Chris, for pointing out the dual mode!

Change-Id: I8522930cebbb9e9ecc732397cbb3deea9e8c9127
Reviewed-on: https://gerrit.libreoffice.org/43122
Reviewed-by: Patrick Jaap <patrick.jaap@tu-dresden.de>
Reviewed-by: Bartosz Kosiorek <gang65@poczta.onet.pl>
Tested-by: Bartosz Kosiorek <gang65@poczta.onet.pl>