tdf#112860 sw: avoid shortcut for tables in SwFrame::GetNextSctLeaf()
As the comment says the shortcut only works if no columns or pages are
between the section frames, while for tables there are definitely table,
row and column frames. This resulted later in
SwFrame::IsFootnoteAllowed() a nullptr dereference, as the frame was
still marked as in-table, but wasn't in a table anymore.

Change-Id: I4191c77d20c6ccc096156fae15ca7c22cc5b68b6
Reviewed-on: https://gerrit.libreoffice.org/43091
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>