EPUB export: better handling of tabs
Mapping ODF/librevenge tab to \t in HTML is not a great idea, as it's
ignorable whitespace. Go with NBSPs and a breakable space instead, that
is much closer visually (15 is just an arbitrary number, it's what MS
Word uses in its HTML export, LO Writer HTML export doesn't handle
this).

Adapt the empty paragraph case to also use NBSP for consistency.

Change-Id: I131802416499eb4f3a83a333b37ca20b59fcd56a