tdf#105381 ww8export: support footnote numbering per page
This is only for the benefit of MSOffice, LO already uses
Dop.rncFootnote for this, but MSO requires this section sprm.
LO import currently ignores sprmSRncFtn.

Endnotes in LO do not currently support .eNum, so we can't
support MSO's rncRstSect for sprmSRncEdn, leaving only
the default - so no use writing that out.

Change-Id: Iaff703e806d1cb62bc5990eb1cdfe323705bf7ee
Reviewed-on: https://gerrit.libreoffice.org/43622
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>