TSCP: don't modify the class. object, delete and create a new one
In Impress advanced classification, a new text object is created
the first time and added to a master page, and then later,
if changed, the object ismodified. The modification however
didn't work, so instead of modifiying, try to delete all text
objects (identified for classification) from the master page and
then create a new one from scratch.

Change-Id: I2263949a1c08e98ce066883648d866175f42c91d
Reviewed-on: https://gerrit.libreoffice.org/44016
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>