Resolves: tdf#111428 implement CellStoreEvent::swap() in ScColumn::SwapCol()
CellStoreEvent remembered the original ScColumn* and corresponding
ScColumn::mnBlkCountFormula, hence after inserting a column to the
left the quick check of ScColumn::HasFormulaCell() whether there
are any formula cells worked on the swapped in column (originally
the one to the right) that happens to be empty in the scenario.
Things worked correctly by accident if the next column to the
right already contained a formula cell.

Regression from

    commit ba61f5b37f455f08cce6699c324abb5a65647791
    Date:   Sat Jul 16 20:01:48 2016 -0400

Change-Id: If993856ceee657736f516a81c293506041a6b7eb